+++ src/ArrayStore.php
@@ -54,7 +54,7 @@ class ArrayStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function get($key)
+    public function getOrFail($key)
     {
         KeyUtil::validate($key);
 
@@ -82,12 +82,12 @@ class ArrayStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function getMultiple(array $keys)
+    public function getMultipleOrFail(array $keys)
     {
         $values = array();
 
         foreach ($keys as $key) {
-            $values[$key] = $this->get($key);
+            $values[$key] = $this->getOrFail($key);
         }
 
         return $values;