+++ src/JsonFileStore.php
@@ -68,7 +68,7 @@ class JsonFileStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function get($key)
+    public function getOrFail($key)
     {
         KeyUtil::validate($key);
 
@@ -112,7 +112,7 @@ class JsonFileStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function getMultiple(array $keys)
+    public function getMultipleOrFail(array $keys)
     {
         $values = array();
         $data = $this->load();