+++ src/PhpRedisStore.php
@@ -73,7 +73,7 @@ class PhpRedisStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function get($key)
+    public function getOrFail($key)
     {
         KeyUtil::validate($key);
 
@@ -117,7 +117,7 @@ class PhpRedisStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function getMultiple(array $keys)
+    public function getMultipleOrFail(array $keys)
     {
         KeyUtil::validateMultiple($keys);