+++ src/RiakStore.php
@@ -73,7 +73,7 @@ class RiakStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function get($key)
+    public function getOrFail($key)
     {
         KeyUtil::validate($key);
 
@@ -115,7 +115,7 @@ class RiakStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function getMultiple(array $keys)
+    public function getMultipleOrFail(array $keys)
     {
         KeyUtil::validateMultiple($keys);